[SC-53871] Implements ability avoid excessive WYU downloads when we have a bad update. (#20)
* [SC-53871] Implements ability avoid excessive WYU downloads when we have a bad update.

The original context here is that at one point in the past we released an agent that was crashed on execution.
This enabled a bad cycle where current working agents in the field would see an available update download it and process it.
hUpdate would download the candidate WYS file and corresponding WYU file, go ahead and "backup" the existing agent service artifacts,
stop the service, install the new update, and try to restart the service.  The service start attempt would fail, and hUpdate would rollback
that install.  At the next agent update attempt, this would happen again.

This lead to huge bloat in S3 traffic and corresponding cost and windows of downtime for the agent as it was processing a failed install.

This change introduces a new sentinel file, i.e. the `this-install-failed.wys`, into the agent's installed folder.  This sentinel file will be
the actual WYS file downloaded to process an update.  The new logic would work as follows:
1. If an agent install fails, then we'll leave the WYS file that drove said update as the sentinel file behind as a breadcrumb.
2. Whenever the agent attempts to do an update, which is triggered by various means, it will always download a candidate WYS file from S3
and check against the sentinel file.  If they are the same, the update is abandoned.  The WYS file currently hovers < 1KiB as oppossed to the WYU file, which is > 50MiB.
3. Otherwise, the update would work as it previously did.

There is one hole here in that if something could be done to keep planting a sentinel file that matches the live "good update" file off S3,
then someone could thwart allowing the agent to update itself; however, there are many more ways to accomplish the same thing with much less effort given our current implementation of hUpdate.

* Applied feedback:
* Used %w vs %v for formatting errors
* Used more efficient form of creating a []string
* Removed use of the term Context and ctx from the CandidateUpdateRequest struct to avoid confusion with the more popluar context structure
* Killed use of ioutils
* Cleaned up consts declaration
* Renamed failed sentinel file